OPS-99: set rbenv user UID in wrapper recipe
In order to have consistent UIDs for the rbenv user across all our
systems, we should set the UID to a specific value via an attribute,
and reopen the user created in the rbenv default recipe. This is to
avoid problems in the future where rbenv snags a UID in our "human
users" range.

* set the rbenv uid as an attribute, 216 was chosen as it wasn't in
  use by any systems (via knife search)
* update test
* update berksfile for berkshelf v3